feat: split #mtglist to another file
